refactor | sprint1 | LC-344 | controller 비즈니스 로직 삭제 | SHUN
- controller에 있는 비즈니스 로직 service로 이동
- controller 및 service 구조 임시 리팩토링
