+++ backend-dashboard/src/main/java/com/backend_dashboard/backend_dashboard/mainPage/controller/MainPageController.java
@@ -24,30 +24,16 @@ public class MainPageController {
 
     @GetMapping(value = "/sensor-stream", produces = MediaType.TEXT_EVENT_STREAM_VALUE)
     public Flux<String> streamSensorData() {
-        final LocalDateTime[] lastFetchTime = {LocalDateTime.now(ZoneOffset.UTC).minusSeconds(2)};
+        final LocalDateTime[] lastFetchTime = {LocalDateTime.now(ZoneOffset.UTC).minusSeconds(3)};
 
-        return Flux.interval(Duration.ofSeconds(1))
+        return Flux.interval(Duration.ofSeconds(2))
                 .flatMap(tick -> {
                     LocalDateTime currentFetchTime = LocalDateTime.now(ZoneOffset.UTC).minusSeconds(1);
-                    return sensorDataService.getRecentSensorData(lastFetchTime[0])
+                    return sensorDataService.getLatestSensorData(lastFetchTime[0])
                             .collectList()
                             .flatMapMany(list -> {
                                 if (!list.isEmpty()) {
-                                    // 데이터 있을 때만 업데이트
                                     lastFetchTime[0] = currentFetchTime;
-
-                                    // sensor_id 기준 최신 데이터 필터링
-                                    Map<String, SensorDataDto> latestPerSensor = new HashMap<>();
-                                    for (SensorDataDto dto : list) {
-                                        String sensorId = dto.getSensor_id();
-                                        SensorDataDto existing = latestPerSensor.get(sensorId);
-                                        // 최신 데이터만 저장 (timestamp 비교)
-                                        if (existing == null || isAfter(dto, existing)) {
-                                            latestPerSensor.put(sensorId, dto);
-                                        }
-                                    }
-                                    // 최신 데이터만 리스트로 반환
-                                    list = new ArrayList<>(latestPerSensor.values());
                                 }
                                 return Flux.just(list);
                             });
@@ -62,17 +48,6 @@ public class MainPageController {
                 })
                 .filter(msg -> !msg.trim().isEmpty())
                 .onErrorResume(e -> Flux.empty());
-    }
-
-    // timestamp 비교 메서드
-    private boolean isAfter(SensorDataDto a, SensorDataDto b) {
-        try {
-            LocalDateTime timeA = LocalDateTime.parse(a.getTimestamp());
-            LocalDateTime timeB = LocalDateTime.parse(b.getTimestamp());
-            return timeA.isAfter(timeB);
-        } catch (Exception e) {
-            return false;
-        }
     }
 
 }
