+++ backend-dashboard/src/main/java/com/backend_dashboard/backend_dashboard/mainPage/service/MainPageService.java
@@ -1,17 +1,18 @@
 package com.backend_dashboard.backend_dashboard.mainPage.service;
 
-import com.backend_dashboard.backend_dashboard.mainPage.domain.dto.SensorDataDto;
-import com.fasterxml.jackson.databind.ObjectMapper;
+import com.backend_dashboard.backend_dashboard.mainPage.domain.dto.ParticleSensorDataDto;
 import lombok.RequiredArgsConstructor;
 import org.elasticsearch.action.search.SearchRequest;
 import org.elasticsearch.client.RequestOptions;
 import org.elasticsearch.client.RestHighLevelClient;
+import com.fasterxml.jackson.databind.ObjectMapper;
 import org.elasticsearch.index.query.QueryBuilders;
 import org.elasticsearch.index.query.RangeQueryBuilder;
 import org.elasticsearch.search.SearchHit;
 import org.elasticsearch.search.builder.SearchSourceBuilder;
 import org.elasticsearch.search.sort.SortOrder;
 import org.springframework.stereotype.Service;
+import com.backend_dashboard.backend_dashboard.mainPage.domain.dto.SensorDataDto;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 import reactor.core.scheduler.Schedulers;
@@ -19,7 +20,9 @@ import reactor.core.scheduler.Schedulers;
 import java.time.LocalDateTime;
 import java.time.ZoneOffset;
 import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
 @Service
 @RequiredArgsConstructor
@@ -28,11 +31,27 @@ public class MainPageService {
     private final RestHighLevelClient client;
     private final ObjectMapper objectMapper = new ObjectMapper();
 
-//    private static final String INDEX = "sensor_data_stream";
     private static final String INDEX = "iot-sensor-data";
 
+    // Controller가 호출하는 메서드
+    public Flux<SensorDataDto> getLatestSensorData(LocalDateTime fromTime) {
+        return getRecentSensorData(fromTime)
+                .collectList()
+                .flatMapMany(list -> {
+                    Map<String, SensorDataDto> latestPerSensor = new HashMap<>();
+                    for (SensorDataDto dto : list) {
+                        String sensorId = dto.getSensor_id();
+                        SensorDataDto existing = latestPerSensor.get(sensorId);
+                        if (existing == null || isAfter(dto, existing)) {
+                            latestPerSensor.put(sensorId, dto);
+                        }
+                    }
+                    return Flux.fromIterable(latestPerSensor.values());
+                });
+    }
 
-    public Flux<SensorDataDto> getRecentSensorData(LocalDateTime fromTime) {
+    // 서비스 내부에서 사용하는 메서드
+    private Flux<SensorDataDto> getRecentSensorData(LocalDateTime fromTime) {
         SearchRequest request = new SearchRequest(INDEX);
 
         RangeQueryBuilder rangeQuery = QueryBuilders
@@ -61,5 +80,15 @@ public class MainPageService {
                     }
                     return Flux.fromIterable(result);
                 });
+    }
+
+    private boolean isAfter(SensorDataDto a, SensorDataDto b) {
+        try {
+            LocalDateTime timeA = LocalDateTime.parse(a.getTimestamp());
+            LocalDateTime timeB = LocalDateTime.parse(b.getTimestamp());
+            return timeA.isAfter(timeB);
+        } catch (Exception e) {
+            return false;
+        }
     }
 }
